ダメージ時無敵時間の実装 色変えの実装 連続攻撃アニメーションがたまに変な状態で止まるバグの修正

+++ Assets/GravityPainter/System/Player/Program/Gpt_PlayerAttackState.cs
@@ -25,9 +25,9 @@ public class Gpt_PlayerAttackState : MonoBehaviour {
 
     void DrawEnemy(Gpt_EnemyColor enemyColor)
     {
-        if (enemyColor.GetColor() != playerState.PlayerColor)
+        if (enemyColor.GetColor() != (int)playerState.PlayerColor)
         {
-            enemyColor.SetColor(playerState.PlayerColor);
+            enemyColor.SetColor((int)playerState.PlayerColor);
             playerState.AddPlayerColorCombo();
         }
     }
